Add travis with deno test
