+++ js-practices/src/queue.ts
@@ -1,7 +1,12 @@
 import {createNode, Node} from "./node.ts";
-import { assertEquals } from "https://deno.land/std/testing/asserts.ts";
 
-function createQueue() {
+export interface Queue{
+    enqueue(v: number): void;
+    dequeue(v: number): number | undefined;
+    toArray(): number[];
+}
+
+export function createQueue(): Queue {
     let root: Node | undefined;
     return {
         enqueue(v: number) {
@@ -39,7 +44,7 @@ function createQueue() {
     }
 }
 
-function createArrayQueue() {
+export function createArrayQueue(): Queue {
     let head = 0;
     let tail = -1;
     let queue: number[] = [];
@@ -70,27 +75,4 @@ function createArrayQueue() {
         }
     }
 }
-
-// tests
-
-function runTests(queueCreator: () => any) {
-    Deno.test(`${queueCreator.name} enqueue should work as expected`, () => {
-        const queue = queueCreator();
-        queue.enqueue(1)
-        queue.enqueue(2)
-        queue.enqueue(3)
-        assertEquals(queue.toArray(), [1,2,3]);
-    });
-
-    Deno.test(`${queueCreator.name} dequeue should follow FIFO`, () => {
-        const queue = queueCreator();
-        queue.enqueue(1)
-        queue.enqueue(2)
-        queue.enqueue(3)
-        assertEquals(queue.dequeue(), 1);
-        assertEquals(queue.toArray(), [2, 3]);
-    });
-}
-
-[createQueue, createArrayQueue].forEach(runTests)
